Support setup.py in the python docker images

+++ docker/docker-python-entrypoint.sh
@@ -2,5 +2,9 @@
 
 virtualenv -p python snyk
 source snyk/bin/activate
-pip install -U -r "${PROJECT_PATH}/requirements.txt"
+if [ -f "${PROJECT_PATH}/requirements.txt" ]; then
+    pip install -U -r "${PROJECT_PATH}/requirements.txt"
+elif [ -f "${PROJECT_PATH}/setup.py" ]; then
+    pip install -U "${PROJECT_PATH}"
+fi
 bash docker-entrypoint.sh "$@"
